% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Py_code.R
\name{find_Py_code}
\alias{find_Py_code}
\title{Search for Python Functions by Descriptions}
\usage{
find_Py_code(goal)
}
\arguments{
\item{goal}{Goal Descriptions}
}
\description{
Search for Python Functions by Descriptions
}
\examples{
find_Py_code("set directory")
}
