% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_code.R
\name{find_R_code}
\alias{find_R_code}
\title{Search for R Functions by Descriptions}
\usage{
find_R_code(goal)
}
\arguments{
\item{goal}{Goal Descriptions}
}
\description{
Search for R Functions by Descriptions
}
\examples{
find_R_code("set directory")
}
