% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyr.R
\name{pyr}
\alias{pyr}
\title{Translate Python Functions into R Functions}
\usage{
pyr(Py_code)
}
\arguments{
\item{Py_code}{Python Code String}
}
\description{
Translate Python Functions into R Functions
}
\examples{
pyr("os.chdir")
}
