% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyryp.R
\docType{package}
\name{pyryp}
\alias{pyryp}
\alias{pyryp-package}
\title{pyryp: a R-Python translator}
\description{
Translator of Common R and Python Functions
}
\author{
Rui K. Yang \email{rkzyang@gmail.com}

Luyao Peng \email{luyaopeng.cn@gmail.com}
}
