% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpy.R
\name{rpy}
\alias{rpy}
\title{Translate R Functions into Python Functions}
\usage{
rpy(R_code)
}
\arguments{
\item{R_code}{R Code String}
}
\description{
Translate R Functions into Python Functions
}
\examples{
rpy("getwd")
}
